+++ embark/embark/settings/deploy.py
@@ -315,6 +315,9 @@ VERSION = get_version_strings()
 # Worker setup
 WORKER_FILES_PATH = os.path.join(BASE_DIR.parent, "WORKER_FILES")
 WORKER_SETUP_LOGS = os.path.join(WORKER_FILES_PATH, "logs/worker_setup_{timestamp}.log")
+WORKER_EMBA_ROOT = "/root/emba/"
+WORKER_FIRMWARE_DIR = "/root/firmware/"
+WORKER_EMBA_LOGS = "/root/emba_logs/"
 
 # Celery task queue
 CELERY_BROKER_URL = f"redis://{REDIS_HOST}:{REDIS_PORT}/0"